+++ lib/Command/Scan.php
@@ -60,7 +60,7 @@ class Scan extends Command {
 
 		if (count($users) === 0) {
 			$output->writeln("<error>Please specify a valid user id to scan, \"--all\" to scan for all users<error>");
-            return Command::FAILURE;
+            return 1;
 		}
 		
 		foreach ($users as $userId) {
@@ -75,6 +75,6 @@ class Scan extends Command {
                 $this->scanner->scanForAudios($userId, $output);
 			}
 		}
-        return Command::SUCCESS;
+        return 0;
     }
 }